% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esRd_estimate_parameters.R
\name{esRd_estimate_parameters}
\alias{esRd_estimate_parameters}
\title{Estimate parameters.}
\usage{
esRd_estimate_parameters(
  clusters_size,
  clusters_freq,
  prior_r = c(10, 10),
  prior_k = c(5, 10),
  mean_generation_interval = 5.2,
  prior_number_yearly_mutations = c(14, 0.5),
  prior_testing = c(1, 3, 0.05, 1),
  sequencing_proba,
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  control = NULL
)
}
\arguments{
\item{clusters_size}{A list of sizes of identical sequence clusters.}

\item{clusters_freq}{A list of frequencies of identical sequence clusters.}

\item{prior_r}{Shape and rate parameter for the prior distribution (gamma) of
the effective reproduction number.}

\item{prior_k}{Shape and rate parameter for the prior distribution (gamma) of
the dispersion parameter.}

\item{mean_generation_interval}{Mean generation interval.}

\item{prior_number_yearly_mutations}{Mean and standard deviation for the prior
distribution (normal) of the number of yearly mutations.}

\item{prior_testing}{Parameters (shape and interval boundaries) for the prior
distribution (scaled beta) of the fraction of cases that are confirmed by a test.}

\item{sequencing_proba}{Sequencing probability.}
}
\value{
An object of S4 class \code{stanfit} containing the fitted results.
}
\description{
\code{esRd_estimate_parameters()} estimates the effective reproduction number,
the dispersion parameter, the probability of a case undergoing a mutation and
the probability that a case is confirmed by a test based on the size distribution
of identical sequence clusters.
}
\details{
The core of the function \code{esRd_estimate_parameters()} is a mathematical model
of the size distribution of sequence clusters, in which viral transmission,
the mutation of the virus, and incomplete case-detection are integrated.
Parameters are estimated by a Bayesian inference model implemented in Stan.
}
\examples{
options(mc.cores = parallel::detectCores())
esRd_estimate_parameters(
  c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 13, 14, 15, 16, 17, 18, 20, 22, 23, 28, 29, 32, 33, 34, 35, 36, 83, 103),
  c(703, 117, 49, 37, 19, 17, 5, 15, 4, 3, 1, 3, 4, 2, 2, 1, 3, 2, 1, 1, 1, 1, 1, 1, 1, 2, 2, 1, 1),
  c(10,10),
  c(5,10),
  5.2,
  c(14,0.5),
  c(1,3,0.05,1),
  1)
}
