% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis-package.R
\docType{package}
\name{estRodis-package}
\alias{estRodis-package}
\alias{estRodis}
\title{The 'estRodis' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.16. https://mc-stan.org
}
