% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_estimate_parameters_1.R
\name{estRodis_estimate_parameters_1}
\alias{estRodis_estimate_parameters_1}
\title{Estimate parameters.}
\usage{
estRodis_estimate_parameters_1(
  clusters_size,
  clusters_freq,
  prior_r = c(10, 10),
  prior_k = c(5, 10),
  mean_generation_interval = 5.2,
  prior_number_yearly_mutations = c(14, 0.5),
  prior_testing = c(1, 3, 0.05, 1),
  sequencing_proba = 1,
  pars = NA,
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  seed = sample.int(.Machine$integer.max, 1),
  init = "random",
  check_data = TRUE,
  sample_file = NULL,
  diagnostic_file = NULL,
  verbose = FALSE,
  algorithm = c("NUTS", "HMC", "Fixed_param"),
  control = NULL,
  include = TRUE,
  cores = getOption("mc.cores", 1L),
  open_progress = interactive() && !isatty(stdout()) && !identical(Sys.getenv("RSTUDIO"),
    "1"),
  show_messages = TRUE
)
}
\arguments{
\item{clusters_size}{A list of sizes of identical sequence clusters.}

\item{clusters_freq}{A list of frequencies of identical sequence clusters.}

\item{prior_r}{Shape and rate parameter for the prior distribution (gamma) of
the effective reproduction number.}

\item{prior_k}{Shape and rate parameter for the prior distribution (gamma) of
the dispersion parameter.}

\item{mean_generation_interval}{Mean generation interval.}

\item{prior_number_yearly_mutations}{Mean and standard deviation for the prior
distribution (normal) of the number of yearly mutations.}

\item{prior_testing}{Parameters (shape and interval boundaries) for the prior
distribution (scaled beta) of the fraction of cases that are confirmed by a test.}

\item{sequencing_proba}{Probability that a case confirmed by a test gets sequenced.}

\item{pars}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{chains}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{iter}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{warmup}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{thin}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{seed}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{init}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{check_data}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{sample_file}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{diagnostic_file}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{verbose}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{algorithm}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{control}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{include}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{cores}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{open_progress}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}

\item{show_messages}{Parameter that is passed on to rstan::sampling, see manual of rstan::sampling for more details.}
}
\value{
An object of S4 class \code{stanfit} containing the fitted results.
}
\description{
\code{estRodis_estimate_parameters_1()} estimates the effective reproduction number,
the dispersion parameter, the probability of a case undergoing a mutation and
the probability that a case is confirmed by a test based on the size distribution
of identical sequence clusters.
}
\details{
The core of the function \code{estRodis_estimate_parameters_1()} is a mathematical model
of the size distribution of sequence clusters, in which viral transmission,
the mutation of the virus, and incomplete case-detection are integrated.
Parameters are estimated by a Bayesian inference model implemented in Stan.
The effective reproduction number, the dispersion parameter, the number of yearly mutations
and the testing probability are included into the model with prior distributions.
The mean generation interval and the sequencing probability are included into
the model as fixed parameters.
More details can be found in the following paper: ADD REFERENCE
}
\examples{
simulated_clusters <- estRodis_simulate_cluster_sizes(n_clusters = 1000,
max_cluster_size = 2500, R = 0.8, k = 0.3, yearly_mutation_rate = 14,
mean_generation_interval = 5.2, testing_proba = 0.6, sequencing_proba = 0.4)

options(mc.cores = parallelly::availableCores())

estRodis_estimate_parameters_1(
  clusters_size = simulated_clusters$size,
  clusters_freq = simulated_clusters$frequency,
  sequencing_proba = 0.44)
}
