% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_init_params_model_five.R
\name{estRodis_init_params_model_five}
\alias{estRodis_init_params_model_five}
\title{Initial values for Model 5}
\usage{
estRodis_init_params_model_five(
  prior_r = c(10, 10),
  prior_k = c(5, 10),
  prior_mutation = c(27, 68),
  prior_testing = c(1, 3, 0.05, 1)
)
}
\arguments{
\item{prior_r}{Parameters for prior distribution of R (shape and rate parameter for gamma distribution)}

\item{prior_k}{Parameters for prior distribution of k (shape and rate parameter for gamma distribution)}

\item{prior_mutation}{Parameters for prior distributions of mutation probability (shape parameters for beta distribution)}

\item{prior_testing}{Parameters for prior distribution of testing probability (scaled beta distribution)}
}
\value{
A named list containing samples of the prior distributions of R, k, the mutation probability and the testing probability.
}
\description{
Sample from prior distributions of R, k, the mutation probability and the testing probability.
}
\examples{
estRodis_init_params_model_five()
}
