% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_init_params_model_four.R
\name{estRodis_init_params_model_four}
\alias{estRodis_init_params_model_four}
\title{Initial values for Model 4}
\usage{
estRodis_init_params_model_four(prior_r = c(10, 10), prior_k = c(5, 10))
}
\arguments{
\item{prior_r}{Parameters for prior distribution of R (shape and rate parameter for gamma distribution)}

\item{prior_k}{Parameters for prior distribution of k (shape and rate parameter for gamma distribution)}
}
\value{
A named list containing samples of the prior distributions of R and k.
}
\description{
Sample from prior distributions of R and k.
}
\examples{
estRodis_init_params_model_four()
}
