% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_init_params_model_one.R
\name{estRodis_init_params_model_one}
\alias{estRodis_init_params_model_one}
\title{Initial values for Model 1}
\usage{
estRodis_init_params_model_one(
  prior_r = c(10, 10),
  prior_k = c(5, 10),
  prior_number_yearly_mutations = c(14, 0.5),
  prior_testing = c(1, 3, 0.05, 1)
)
}
\arguments{
\item{prior_r}{Parameters for prior distribution of R (shape and rate parameter for gamma distribution)}

\item{prior_k}{Parameters for prior distribution of k (shape and rate parameter for gamma distribution)}

\item{prior_number_yearly_mutations}{Parameters for prior distribution of the number of yearly mutations (mean and standard deviation for normal distribution)}

\item{prior_testing}{Parameters for prior distribution of testing probability (scaled beta distribution)}
}
\value{
A named list containing samples of the prior distributions of R, k, the number of yearly mutations and the testing probability.
}
\description{
Sample from prior distributions of R, k, the number of yearly mutations and the testing probability.
}
\examples{
estRodis_init_params_model_one()
}
