% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_init_params_model_six.R
\name{estRodis_init_params_model_six}
\alias{estRodis_init_params_model_six}
\title{Initial values for Model 6}
\usage{
estRodis_init_params_model_six(
  prior_r = c(10, 10),
  prior_k = c(5, 10),
  prior_mutation = c(27, 68)
)
}
\arguments{
\item{prior_r}{Parameters for prior distribution of R (shape and rate parameter for gamma distribution)}

\item{prior_k}{Parameters for prior distribution of k (shape and rate parameter for gamma distribution)}

\item{prior_mutation}{Parameters for prior distributions of mutation probability (shape parameters for beta distribution)}
}
\value{
A named list containing samples of the prior distributions of R, k and the mutation probability.
}
\description{
Sample from prior distributions of R, k and the mutation probability.
}
\examples{
estRodis_init_params_model_six()
}
