% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_init_params_model_three.R
\name{estRodis_init_params_model_three}
\alias{estRodis_init_params_model_three}
\title{Initial values for Model 3}
\usage{
estRodis_init_params_model_three(
  prior_r = c(10, 10),
  prior_k = c(5, 10),
  prior_testing = c(1, 3, 0.05, 1)
)
}
\arguments{
\item{prior_r}{Parameters for prior distribution of R (shape and rate parameter for gamma distribution)}

\item{prior_k}{Parameters for prior distribution of k (shape and rate parameter for gamma distribution)}

\item{prior_testing}{Parameters for prior distribution of testing probability (scaled beta distribution)}
}
\value{
A named list containing samples of the prior distributions of R, k and the testing probability.
}
\description{
Sample from prior distributions of R, k and the testing probability.
}
\examples{
estRodis_init_params_model_three()
}
