% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_init_params_model_two.R
\name{estRodis_init_params_model_two}
\alias{estRodis_init_params_model_two}
\title{Initial values for Model 2}
\usage{
estRodis_init_params_model_two(
  prior_r = c(10, 10),
  prior_k = c(5, 10),
  prior_number_yearly_mutations = c(14, 0.5)
)
}
\arguments{
\item{prior_r}{Parameters for prior distribution of R (gamma)}

\item{prior_k}{Parameters for prior distribution of k (gamma)}

\item{prior_number_yearly_mutations}{Parameters for prior distribution of the number of yearly mutations (normal)}
}
\value{
A named list containing samples of the prior distributions of R, k and the number of yearly mutations.
}
\description{
Sample from prior distributions of R, k, the number of yearly mutations and the testing probability.
}
\examples{
estRodis_init_params_model_two()
}
