% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_plot_transmission_chain_mutation.R
\name{estRodis_plot_transmission_chain_mutation}
\alias{estRodis_plot_transmission_chain_mutation}
\title{Plot transmission chain with mutation}
\usage{
estRodis_plot_transmission_chain_mutation(
  transmission_chain_nodes,
  transmission_chain_edges,
  max_generation = max(dplyr::pull(transmission_chain_nodes, "generation")),
  style_plot = "flexible",
  style_legend_clusters = "flexible"
)
}
\arguments{
\item{transmission_chain_nodes}{Data frame containing information about the nodes of the transmission chain}

\item{transmission_chain_edges}{Data frame containing information about the edges of the transmission chain}

\item{max_generation}{Maximum generation of nodes and edges that should be included into the plot}

\item{style_plot}{Positioning of transmission chain on canvas}

\item{style_legend_clusters}{Level of detail of the legend of the identical sequence clusters}
}
\value{
Plot of the transmission chain defined by nodes, edges and max_generation in the form of a cowplot plot_grid object.
}
\description{
A plot of a transmission chain including the mutation process is created.
}
\details{
The data frame \code{transmission_chain_nodes} has to contain at least the columns \code{node_key} (id of node),
\code{generation} (generation of nodes), \code{mutation_occurred} (whether or not a mutation occurred),
\code{variant_received} (variant received from parent node),
\code{variant_after_mutation} (variant present at node in case a mutation occurs) and
\code{current_variant} (variant of virus present at node at time of detection).

The data frame \code{transmission_chain_edges} has to contain at least the columns
\code{from} (index of infector node in \code{transmission_chain_nodes}),
\code{to} (index of infectee node in \code{transmission_chain_nodes}),
\code{generation} (generation of infectee node),
\code{from_variant} (variant of virus present at infector node at time of detection),
\code{to_variant} (variant of virus present at infectee node at time of detection), and
\code{variant_transmitted} (variant transmitted from infector to infectee).

Using the inputs \code{max_generation}, \code{style_plot} and \code{style_legend_clusters},
the graphical presentation of the transmission chain can be influenced.

\code{max_generation} determines the last generation of nodes and edges
that are included into the plot.

Setting \code{style_plot = fixed}, nodes and edges up to (and including)
generation \code{max_generation} are drawn,
while nodes and edges of later generations are made invisible.

Setting \code{style_plot = flexible}, nodes and edges up to (and including)
generation \code{max_generation} are drawn,
while nodes and edges of later generations are cut away.

Setting \code{style_legend_clusters = fixed}, the legend of the
identical sequence clusters includes all identical sequence clusters
of the transmission chain defined by nodes and edges regardless of their generation.

Setting \code{style_legend_clusters = flexible}, the legend of the
identical sequence clusters only includes all identical sequence clusters
of the transmission chain defined by nodes and edges up to (and including)
generation \code{max_generation}.
}
