% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_scaled_beta_distribution.R
\name{estRodis_scaled_beta_distribution_pdf}
\alias{estRodis_scaled_beta_distribution_pdf}
\title{Probability density function of scaled Beta distribution}
\usage{
estRodis_scaled_beta_distribution_pdf(alpha, beta, p, q, x)
}
\arguments{
\item{alpha}{shape parameter (real number > 0)}

\item{beta}{shape parameter (real number > 0)}

\item{p}{lower boundary of interval on which the distribution is defined}

\item{q}{upper boundary of interval on which the distribution is defined}

\item{x}{list of values at which the probability density function of the scaled beta distribution shall be evaluated}
}
\value{
A list of real numbers.
}
\description{
\code{estRodis_scaled_beta_distribution_pdf} evaluates the probability density function of the scaled Beta distribution that is
determined by the parameters \code{alpha} and \code{beta} and defined on the interval  \code{p}, \code{q} at \code{x}.
}
\details{
Values of \code{x} outside the interval \code{p}, \code{q} are evaluated to \code{0}.
}
