% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_simulate_cluster_sizes.R
\name{estRodis_simulate_cluster_sizes}
\alias{estRodis_simulate_cluster_sizes}
\title{Simulation of identical sequence clusters.}
\usage{
estRodis_simulate_cluster_sizes(
  n_clusters = 1000,
  max_cluster_size = 2500,
  R = 1,
  k = 0.3,
  yearly_mutation_rate = 14,
  mean_generation_interval = 5.2,
  testing_proba = 0.6,
  sequencing_proba = 0.4
)
}
\arguments{
\item{n_clusters}{Number of clusters to be simulated.}

\item{max_cluster_size}{Maximal size of a simulated identical sequence cluster.}

\item{R}{Effective reproduction number.}

\item{k}{Dispersion parameter.}

\item{yearly_mutation_rate}{Number of mutations of the viral genome per year.}

\item{mean_generation_interval}{Mean generation interval.}

\item{testing_proba}{Probability that a case is confirmed by a test.}

\item{sequencing_proba}{Probability that the genome of a confirmed case is sequenced.}
}
\value{
A \code{data.frame} with three columns, \code{size}, \code{frequency} and \code{percentage},
containing size and frequency of the simulated identical sequence clusters as well as the
proportion of the identical sequence clusters of each size.
}
\description{
\code{n_clusters} identical sequence clusters of sizes \code{1} to \code{max_cluster_size}
are simulated. Viral transmission (\code{R} and \code{k}), viral mutation
(\code{yearly_mutation_rate} and \code{mean_generation_interval}) and case detection
(\code{testing_proba} and \code{sequencing_proba}) are included in the simulation.
}
\details{
More details can be found in the following paper: ADD REFERENCE
}
\examples{
estRodis_simulate_cluster_sizes()
}
