% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_simulate_transmission_chain.R
\name{estRodis_simulate_transmission_chain}
\alias{estRodis_simulate_transmission_chain}
\title{Simulate transmission chain}
\usage{
estRodis_simulate_transmission_chain(
  max_chain_size = 100,
  R = 1,
  k = 0.3,
  yearly_mutation_rate = 14,
  mean_generation_interval = 5.2,
  testing_proba = 0.6,
  sequencing_proba = 0.4
)
}
\arguments{
\item{max_chain_size}{Maximal size of the simulated transmission chain.}

\item{R}{Effective reproduction number.}

\item{k}{Dispersion parameter.}

\item{yearly_mutation_rate}{Number of mutations of the viral genome per year.}

\item{mean_generation_interval}{Mean generation interval.}

\item{testing_proba}{Probability that a case is confirmed by a test.}

\item{sequencing_proba}{Probability that the genome of a confirmed case is sequenced.}
}
\value{
A list containing three data frames, \code{nodes} (information about cases), \code{edges} (information about transmission events)
and \code{free_leaves} (information about cases whose offspring has not been determined yet).

The data frame \code{nodes} contains the columns \code{node_key} (id of node), \code{generation} (generation of nodes),
\code{mutation_occurred} (whether or not a mutation occurred), \code{variant_received} (variant received from parent node),
\code{variant_after_mutation} (variant present at node in case a mutation occurs),
\code{current_variant} (variant of virus present at node at time of detection) and \code{detection} (whether or not a case is detected).

The data frame \code{edges} contains the columns \code{from} (index of infector node in \code{nodes}), \code{to} (index of infectee node in \code{nodes}),
\code{generation} (generation of infectee node), \code{from_variant} (variant of virus present at infector node at time of detection),
\code{to_variant} (variant of virus present at infectee node at time of detection), \code{variant_transmitted} (variant transmitted from infector to infectee),
\code{from_detected} (whether or not the infector node has been detected) and \code{to_detected} (whether or not the infectee node has been detected).

The data frame \code{free_leaves} contains the columns \code{node_key} (id of node), \code{generation} (generation of nodes),
\code{mutation_occurred} (whether or not a mutation occurred), \code{variant_received} (variant received from parent node) and
\code{variant_after_mutation} (variant present after mutation at node in case a mutation occurs at that node).
}
\description{
A transmission chain of maximal size \code{max_chain_size} is simulated.

Viral transmission (\code{R} and \code{k}), viral mutation
(\code{yearly_mutation_rate} and \code{mean_generation_interval}) and case detection
(\code{testing_proba} and \code{sequencing_proba}) are included in the simulation.
}
\examples{
estRodis_simulate_transmission_chain()
}
