% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_plot_transmission_chain.R
\name{estRodis_plot_transmission_chain}
\alias{estRodis_plot_transmission_chain}
\title{Plot transmission chain}
\usage{
estRodis_plot_transmission_chain(
  transmission_chain_nodes,
  transmission_chain_edges,
  max_generation = max(dplyr::pull(transmission_chain_nodes, "generation")),
  colour_nodes_edges = "navyblue",
  style_plot = "flexible",
  width_edges = 1
)
}
\arguments{
\item{transmission_chain_nodes}{Data frame containing information about the nodes of the transmission chain}

\item{transmission_chain_edges}{Data frame containing information about the edges of the transmission chain}

\item{max_generation}{Maximum generation of nodes and edges that are included into the plot}

\item{colour_nodes_edges}{Colour for nodes and edges}

\item{style_plot}{Positioning of transmission chain on canvas}

\item{width_edges}{Width of edges of graph}
}
\value{
Plot of the transmission chain defined by nodes, edges and max_generation in the form of a ggplot2 object.
}
\description{
A plot of a transmission chain is created.
}
\details{
The data frame \code{transmission_chain_nodes} has to contain at least the columns \code{node_key} (id of node)
and \code{generation} (generation of nodes).

The data frame \code{transmission_chain_edges} has to contain at least the columns \code{from}
(index of infector node in \code{transmission_chain_nodes}),
\code{to} (index of infectee node in \code{transmission_chain_nodes})
and \code{generation} (generation of infectee node).

Using the inputs \code{max_generation} and \code{style_plot},
the graphical presentation of the transmission chain can be influenced.

\code{max_generation} determines the last generation of nodes and edges
that are included into the plot.

Setting \code{style_plot = fixed}, nodes and edges up to (and including)
generation \code{max_generation} are drawn,
while nodes and edges of later generations are made invisible.

Setting \code{style_plot = flexible}, nodes and edges up to (and including)
generation \code{max_generation} are drawn,
while nodes and edges of later generations are cut away.
}
