% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRodis_plot_transmission_chain_mutation.R
\name{estRodis_plot_transmission_chain_mutation}
\alias{estRodis_plot_transmission_chain_mutation}
\title{Plot transmission chain with mutation}
\usage{
estRodis_plot_transmission_chain_mutation(
  transmission_chain_nodes,
  transmission_chain_edges,
  max_generation = max(dplyr::pull(transmission_chain_nodes, "generation")),
  style_plot = "flexible",
  style_legend_clusters = "flexible"
)
}
\arguments{
\item{transmission_chain_nodes}{Data frame containing information about the nodes of the transmission chain}

\item{transmission_chain_edges}{Data frame containing information about the edges of the transmission chain}

\item{max_generation}{Maximum generation of nodes and edges that should be included into the plot}

\item{style_plot}{Positioning of transmission chain on canvas}

\item{style_legend_clusters}{Level of detail of the legend of the identical sequence clusters}
}
\value{
Plot of the transmission chain defined by nodes, edges and max_generation in the form of a cowplot plot_grid object.
}
\description{
Plot transmission chain with mutation
}
